cpufreq: cpu-boost: fixes for underclocking cases
We don't want sync and input_boost to set a higher min than
the max frequency, which can happen if the user underclocks below
the thresholds - so fix it.

Also when using load_based_syncs (which it's set to enabled on FK)
we should be evaluating the load based on max and not the absolute
max (cpuinfo.max_freq) because users can and will underclock and we
want to maintain the correct formula ratio.

Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>